refactor: réduction générale mois par mois
